Only generate above-Uni warning for \p{}, \P{}
This warning is generated when a non-Unicode code point is matched
against a Unicode property in a regular expression.  The warning exists
because Unicode properties are defined only on Unicode code points, and,
subtleties occur for matching them against non-Unicode code points.

Unfortunately, the warning was being generated under circumstances other
than this matching, because it was checked at a common place to both
property matching, and other operations.  The solution is to move the
check for this situation out of the common code, and into the code where
just \p{} and \P{} are handled.

Currently, almost all the code that handles \p{} and \P{} is in ANYOF
nodes.  So that is where the check is moved.  However, a very few
Unicode properties match just a single code point, and the ANYOF nodes
for those are optimized into EXACT nodes.  If such a node is matched
against a non-Unicode code point, there is no warning generated.  (This
bug, [perl #114148], exists with or without this patch.)  One way to fix
it is to not do the optimization, but I think that the gain from the
optimization is far greater than the loss of a warning on this very rare
condition.